add foods to seed file

+++ db/migrate/20190213230249_add_time_to_reaction_entries.rb
@@ -1,7 +1,7 @@
 class AddTimeToReactionEntries < ActiveRecord::Migration[5.1]
   def change
     change_table :reaction_entries do |t|
-      t.integer :time
+      t.bigint :time
     end
   end
 end
